fuse: keep sync read-ahead as sync/foreground for uring
Typically background/async page cache requests are better served
on different cores - one core submits requests and handles completed
pages when available and another core handles (ring) requests
async. So far synchrononous read-ahead was handled async and
is slower and not as well scaling as it could if the request
would be handled on the same queue/core.

Signed-off-by: Bernd Schubert <bschubert@ddn.com>