fuse request wait: disable cpu migration
This is really a hack, but rather efficient. We need it to solve
linux scheduler issues to keep the submitting application on the
same core.

Recent kernels have __wake_up_on_current_cpu() and usage would
not be a hack, but it is not exported - not accessible.
We first need to get fuse-uring merged, to be able to export
and use it.